requirements.txt
awscli==1.22.10
boto3==1.20.11
botocore==1.23.11
colorama==0.4.3
cycler==0.11.0
docutils==0.15.2
fonttools==4.28.3
jmespath==0.10.0
kafka-python==2.0.2
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
py4j==0.10.9
pyasn1==0.4.8
pyparsing==3.0.6
pyspark==3.1.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
urllib3==1.26.7
